docs: update project_knights_travails.md with graph article
The Khan Academy's article about graphs was very theoretical. I found other sources of information that described similar concepts differently to be helpful. The Geeks for Geeks link was particularly useful to understand that I had already learned about graphs because trees and linked lists are forms of graphs. This helped boost my confidence in the Knights Travail project.